Moved links logic to uiState
